refactor: move CommitInspectionProvider EP and implementation under api package
