+++ src/main/kotlin/com/github/lppedd/cc/api/CommitInspectionProvider.kt
@@ -1,5 +1,6 @@
-package com.github.lppedd.cc.inspection
+package com.github.lppedd.cc.api
 
+import com.github.lppedd.cc.inspection.CommitBaseInspection
 import com.intellij.openapi.extensions.ExtensionPointName
 
 internal val INSPECTION_EP = ExtensionPointName<CommitInspectionProvider>(
