+++ src/main/resources/META-INF/plugin.xml
@@ -101,7 +101,7 @@
     />
     <extensionPoint
         name="commitInspectionProvider"
-        interface="com.github.lppedd.cc.inspection.CommitInspectionProvider"
+        interface="com.github.lppedd.cc.api.CommitInspectionProvider"
     />
   </extensionPoints>
 
@@ -109,9 +109,6 @@
     <commitTypeProvider order="first" implementation="com.github.lppedd.cc.api.DefaultCommitTokenProvider" />
     <commitScopeProvider order="first" implementation="com.github.lppedd.cc.api.DefaultCommitTokenProvider" />
     <commitFooterProvider order="first" implementation="com.github.lppedd.cc.api.DefaultCommitTokenProvider" />
-    <commitInspectionProvider
-        order="first"
-        implementation="com.github.lppedd.cc.inspection.DefaultInspectionProvider"
-    />
+    <commitInspectionProvider order="first" implementation="com.github.lppedd.cc.api.DefaultInspectionProvider" />
   </extensions>
 </idea-plugin>
